+++ ContentPatcher/release-notes.md
@@ -5,6 +5,7 @@
 * Added new tokens:
   * `HasWalletItem`: the [special items in the player wallet](https://stardewvalleywiki.com/Wallet).
   * `SkillLevel`: the player's level for a given skill.
+* Added support for conditional map edits in SMAPI 2.9.
 * Added `Wind` value for `Weather` token.
 * Added support for matching subkey/value pairs for tokens which previously required subkeys.
 * Fixed language token always marked 'not valid in this context'.
@@ -13,7 +14,7 @@
 * Fixed `patch summary` order for token subkeys and values.
 
 **Breaking changes:**
-* The `Weather` token's `Sun` value has been split into `Sun` and `Wind`. See the readme for more info.
+* The `Weather` token will now return `Wind` on windy days instead of `Sun`. See the readme for more info.
 
 ## 1.5.3
 * Added `patch summary` hint if `Target` value incorrectly includes a file extension.